chore(retake-ui): split build_retake_controls into focused module
Codex P3 review caught that adding `build_retake_controls()` to
`generation_tab_secondary_controls.py` pushed it to 208 LOC, exceeding
the AGENTS.md 200 LOC hard cap. Move the new builder into its own
focused module (`generation_tab_retake_controls.py`) — matches the
existing one-builder-per-file pattern used by `simple_controls`,
`source_controls`, `generate_controls`, etc. Update the section
orchestrator's import.

After split: 174 LOC (secondary) + 45 LOC (retake), both well under cap.
